Code style change and cleanup for alloc/free macros
Code style:
  1. var = RAPTOR_CALLOC(type, count, size)
  Prefering:
    var = RAPTOR_CALLOC(type, 1, sizeof(*var))
  when count = 1

  2. var = RAPTOR_MALLOC(type, size)

  3. RAPTOR_FREE(type, var)

The consequence here is allocs that mostly fit into 1 line without so
much boilerplate and duplication of types.

The RAPTOR_MALLOC and RAPTOR_CALLOC now do the cast to the return type.

RAPTOR_FREE takes the object type too but always casts arg to void
This certainly contains many wrong types to the arg but might be used
later in some kind of smart type-aware debugging allocator.