Fix some configure portability issues
Fixed a typo in the STRCASECMP conditional

Set CPPFLAGS whether or not the check for xsltSaveResultToString
succeeds, because AC_CHECK_HEADERS() depends on this and the result
of the header check should not depend on the result of the lib check.

Put -DRAPTOR_DEBUG and -DMAINTAINER_MODE into the config header
instead of CPPFLAGS. This prevents the flags from potentially being
passed to "make distcheck" builds, where they can cause trouble
(e.g. absence of git-version.h header)

Removed -g from maintainer CPPFLAGS since it does not belong there
and add to CFLAGS when --enable-debug is given.