Fixes for feeds testing for out of tree builds and portability
Do not use a variable containing a list of filenames both in a
target's dependencies and in the target body, at least outside of
maintainer mode.  Make on Solaris 8 has a funny idea of how to
implement VPATH; it modifies the value of the variable by prefixing
the srcdir path to each filename therein, and this breaks our rule
code.

Remove the use of $? as this is not well-supported outside of GNU
Make and the semantics of it are not needed or wanted.

Adjust check-validate-atom target to look for input fields in either
builddir or srcdir since some are generated and some are source.