Delete win32/ subdirectory
The MSVC6 and MSVC7.1 project files in win32/ fail to build current
versions of Raptor, require external projects for Curl and Expat which do
not appear to be publicly available, are unlikely to get the extensive work
they need in the near future due to John Barstow's busy schedule, and are
redundant now that Raptor can be built on Windows using CMake. Thus, out
they go.

win32/README.md: Moved up to README-cmake.md, as this information is
useful, and CMake build support is not limited to Win32