Call Bison with the real file names
Give Bison the real output file names, so that the implementation file
is not fooled when it wants to find the generated header.  To this
end, generate the real output files, but postprocess them, instead of
creating temporary files with wrong names, and then having Perl fix
the wrong names.