CMakeLists.txt: default to shared libraries by default
raptor2 provides a library and depends on a few other libraries
on it's own (like libxml2).

Some downstream projects are not very careful to include all
the needed libraries from .pc files and use just raptor2 one.

This sometimes causes linkage failures like `guitarix`:
https://hydra.nixos.org/log/4hznma7pl7wvjdhm3x2lr84zpj07s2ax-guitarix-0.42.1.drv
In this case linkage failure happens via underlinked `liblrdf.so`.

While it's a bug in `liblrdf.so` I think it would be reasonable
to enable shared libraries by default for use in distributions as is.

This change enables shared library bu default.