Update turtle_parser->consumed on the last chunk
When it's not the last chunk, any data that hasn't been consumed is
moved to the start of the buffer, and ->consumed is set to the length.
However, ->consumed wasn't being reset for the last chunk - so if you
reused a turtle parser after feeding a last chunk through it, the state
would be inconsistent: the next chunk would get copied after the data
from the previous chunk, but the parser would read from the start of the
buffer...

To fix this, just set ->consumed to 0 after the last chunk, so the
buffer gets cleared.

This fixes a failure in librdf's rdf_parser test, which feeds several
strings through the same parser: the "Adding turtle counted string" test
produces "librdf error XXX - syntax error at XXX" with XXX being binary
garbage as the parser ran off the end of the buffer.